+++ src/simmate/toolkit/structure_prediction/evolution/database/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from .evolutionary_search import EvolutionarySearch
 from .structure_source import StructureSource
+from .evolutionary_search import EvolutionarySearch