+++ src/simmate/toolkit/structure_prediction/evolution/workflows/new_individual.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from pathlib import Path
+import logging
 
 from simmate.workflow_engine import Workflow
 from simmate.database.workflow_results import EvolutionarySearch
@@ -13,11 +15,22 @@ class StructurePrediction__Python__NewIndividual(Workflow):
     def run_config(
         search_id: int,
         structure_source_id: int,
+        directory: Path,
         **kwargs,
     ):
 
         search_db = EvolutionarySearch.objects.get(id=search_id)
         source_db = search_db.structure_sources.get(id=structure_source_id)
+
+        # Check the stop condition of the search and see if this new individual
+        # is even needed. This will catch when a search ends while a new
+        # individual was waiting in the queue
+        if search_db.check_stop_condition():
+            logging.info(
+                "The search ended while this individual was in the queue. "
+                "Canceling new individual."
+            )
+            return
 
         if source_db.is_transformation:
 
@@ -52,6 +65,10 @@ class StructurePrediction__Python__NewIndividual(Workflow):
             state = search_db.subworkflow.run(
                 structure=new_structure,
                 source=source,
+                directory=directory,
                 **search_db.subworkflow_kwargs,
             )
             result = state.result()
+            # NOTE: we tell the workflow to use the same directory. There is
+            # good chance the user indicates that they want to compress the
+            # folder to.