+++ CHANGELOG.md
@@ -29,12 +29,15 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 - add `simmate utilities` command group with `archive-old-runs`
 - add `start-cluster` command for starting many local workers
 - add `structure-prediction` workflows
+- add plotting/output utilities to `EvolutionarySearch` and `relaxation.vasp.staged`
 
 
 **Refactors**
 - evolutionary search now delay creations, transformations, and validation until runtime (used to be at time of structure submission)
 - `directory`, `compress_ouput`, and `run_id` are now default input parameters for subclasses of `Workflow`. If these are unused, the `run_config` must include `**kwargs`
 
+**Fixes**
+- fixed when `source` is not being registered by several workflows
 
 
 # v0.8.0 (2022.08.11)